vecstore: move Txn code into separate file
This is a purely mechanical commit that moves Txn code into a new file.

Epic: CRDB-42943

Release note: None
